# services/catalog-service/src/main/resources/openapi/products-api.yaml
openapi: 3.0.3
info:
  title: Catalog Service - Products API
  description: |
    CRUD API для каталога, заказов и промокодов.
    Ролевая модель: USER (покупатель), SELLER (продавец), ADMIN (полный доступ).
    JWT в заголовке Authorization Bearer, в токене — claim `role`.
  version: 1.0.0

# Глобальная схема: все операции по умолчанию требуют JWT (в Swagger UI появится кнопка Authorize).
security:
  - bearerAuth: []

paths:
  /products:
    post:
      summary: Создание товара
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error_code: VALIDATION_ERROR
                message: Ошибка валидации входных данных
                details:
                  fields:
                    - field: name
                      violation: must not be blank
    get:
      summary: Список товаров с пагинацией и фильтрацией
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Размер страницы
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          description: Фильтр по категории (точное совпадение)
          schema:
            type: string
      responses:
        '200':
          description: Страница товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'

  /products/export:
    get:
      summary: Выгрузка списка товаров в CSV-файл
      operationId: exportProducts
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          description: Фильтр по категории (точное совпадение)
          schema:
            type: string
      responses:
        '200':
          description: CSV-файл со списком товаров
          content:
            text/csv:
              schema:
                type: string
                format: binary

  /orders:
    post:
      summary: Создание заказа
      description: "Роли: USER и ADMIN — разрешено; SELLER — запрещено (403 ACCESS_DENIED)."
      operationId: createOrder
      tags:
        - orders
      parameters:
        - name: X-User-Id
          in: header
          required: false
          description: Идентификатор пользователя (владельца заказа); при Bearer JWT подставляется из токена
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          description: ACCESS_DENIED — роль SELLER не может создавать заказы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: Ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: PRODUCT_INACTIVE или ORDER_HAS_ACTIVE или INSUFFICIENT_STOCK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: PROMO_CODE_INVALID или PROMO_CODE_MIN_AMOUNT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          description: ORDER_LIMIT_EXCEEDED — превышен лимит частоты создания заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /orders/{id}:
    get:
      summary: Получение заказа по ID
      description: "Роли: USER — только свой заказ; SELLER — запрещено; ADMIN — любой заказ."
      operationId: getOrder
      tags:
        - orders
      parameters:
        - name: X-User-Id
          in: header
          required: false
          description: Идентификатор пользователя (владельца заказа); при Bearer JWT подставляется из токена
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заказ найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          description: ACCESS_DENIED (недостаточно прав) или ORDER_OWNERSHIP_VIOLATION
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: ORDER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      summary: Обновление заказа (новый состав позиций)
      description: "Роли: USER — только свой; SELLER — запрещено; ADMIN — любой."
      operationId: updateOrder
      tags:
        - orders
      parameters:
        - name: X-User-Id
          in: header
          required: false
          description: Идентификатор пользователя (владельца заказа); при Bearer JWT подставляется из токена
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Идентификатор заказа
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Заказ обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: ORDER_OWNERSHIP_VIOLATION — заказ принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: ORDER_NOT_FOUND — заказ не найден по ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: INVALID_STATE_TRANSITION, PRODUCT_INACTIVE или INSUFFICIENT_STOCK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: PROMO_CODE_MIN_AMOUNT (при снятии промо после пересчёта)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          description: ORDER_LIMIT_EXCEEDED — превышен лимит частоты обновления заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /orders/{id}/cancel:
    post:
      summary: Отмена заказа
      operationId: cancelOrder
      tags:
        - orders
      parameters:
        - name: X-User-Id
          in: header
          required: false
          description: Идентификатор пользователя (владельца заказа); при Bearer JWT подставляется из токена
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Идентификатор заказа
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заказ отменён (status = CANCELED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          description: ORDER_OWNERSHIP_VIOLATION — заказ принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: ORDER_NOT_FOUND — заказ не найден по ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: INVALID_STATE_TRANSITION — отмена разрешена только из CREATED или PAYMENT_PENDING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /promo-codes:
    post:
      summary: Создание промокода
      description: "Роли: USER — запрещено; SELLER и ADMIN — разрешено."
      operationId: createPromoCode
      tags:
        - promo-codes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: Промокод создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: ACCESS_DENIED — недостаточно прав (только SELLER и ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /products/{id}:
    get:
      summary: Получение товара по ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Товар найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Товар не найден по ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error_code: PRODUCT_NOT_FOUND
                message: Товар не найден
    put:
      summary: Обновление товара
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Товар не найден по ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error_code: PRODUCT_NOT_FOUND
                message: Товар не найден
    delete:
      summary: Мягкое удаление (перевод в статус ARCHIVED)
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Товар переведён в ARCHIVED
        '404':
          description: Товар не найден по ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error_code: PRODUCT_NOT_FOUND
                message: Товар не найден

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. В payload — claim `role` (USER, SELLER, ADMIN).
  schemas:
    ApiError:
      type: object
      required:
        - error_code
        - message
      description: Единый формат ошибки API
      properties:
        error_code:
          type: string
          description: Машиночитаемый код ошибки
          enum:
            - PRODUCT_NOT_FOUND
            - PRODUCT_INACTIVE
            - ORDER_NOT_FOUND
            - ORDER_LIMIT_EXCEEDED
            - ORDER_HAS_ACTIVE
            - INVALID_STATE_TRANSITION
            - INSUFFICIENT_STOCK
            - PROMO_CODE_INVALID
            - PROMO_CODE_MIN_AMOUNT
            - ORDER_OWNERSHIP_VIOLATION
            - VALIDATION_ERROR
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - REFRESH_TOKEN_INVALID
            - ACCESS_DENIED
        message:
          type: string
          description: Человекочитаемое описание
        details:
          type: object
          nullable: true
          description: Дополнительные данные (например, список полей с ошибками валидации)

    ValidationErrorDetails:
      type: object
      description: Детали ошибки валидации (в details при error_code VALIDATION_ERROR)
      properties:
        fields:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              violation:
                type: string

    ProductStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ARCHIVED
      description: Состояние товара

    Product:
      type: object
      description: Сущность товара. Поля используются в ProductCreate, ProductUpdate, ProductResponse.
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор товара (только в response)
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название товара
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Описание товара
        price:
          type: number
          format: decimal
          minimum: 0.01
          description: Цена товара (> 0)
        stock:
          type: integer
          minimum: 0
          description: Остаток на складе (>= 0)
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Категория товара
        status:
          $ref: '#/components/schemas/ProductStatus'
        createdAt:
          type: string
          format: date-time
          description: Время создания (только в response)
        updatedAt:
          type: string
          format: date-time
          description: Время обновления (только в response)

    ProductCreate:
      type: object
      required:
        - name
        - price
        - stock
        - category
        - status
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название товара
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Описание товара
        price:
          type: number
          format: decimal
          minimum: 0.01
          description: Цена товара (> 0)
        stock:
          type: integer
          minimum: 0
          description: Остаток на складе
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Категория товара
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название товара
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Описание товара
        price:
          type: number
          format: decimal
          minimum: 0.01
          description: Цена товара (> 0)
        stock:
          type: integer
          minimum: 0
          description: Остаток на складе
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Категория товара
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - category
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор товара
        name:
          type: string
          description: Название товара
        description:
          type: string
          nullable: true
          description: Описание товара
        price:
          type: number
          format: decimal
          description: Цена товара
        stock:
          type: integer
          description: Остаток на складе
        category:
          type: string
          description: Категория товара
        status:
          $ref: '#/components/schemas/ProductStatus'
        createdAt:
          type: string
          format: date-time
          description: Время создания
        updatedAt:
          type: string
          format: date-time
          description: Время обновления

    ProductPage:
      type: object
      required:
        - content
        - totalElements
        - page
        - size
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
          description: Список товаров на странице
        totalElements:
          type: integer
          description: Общее количество элементов
        page:
          type: integer
          description: Номер текущей страницы
        size:
          type: integer
          description: Размер страницы

    OrderStatus:
      type: string
      enum:
        - CREATED
        - PAYMENT_PENDING
        - PAID
        - SHIPPED
        - COMPLETED
        - CANCELED
      description: Статус заказа

    OrderItemCreate:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
          description: Идентификатор товара
        quantity:
          type: integer
          minimum: 1
          maximum: 999
          description: Количество единиц

    OrderCreate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemCreate'
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true
          description: Код промо (опционально)

    OrderUpdate:
      type: object
      required:
        - items
      description: Новый состав позиций заказа (при обновлении разрешён только в статусе CREATED)
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemCreate'

    OrderItemResponse:
      type: object
      required:
        - product_id
        - quantity
        - price_at_order
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_order:
          type: number
          format: decimal

    OrderResponse:
      type: object
      required:
        - id
        - user_id
        - status
        - total_amount
        - discount_amount
        - items
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        total_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PromoCodeCreate:
      type: object
      required:
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - valid_from
        - valid_until
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          description: Уникальный код промокода
        discount_type:
          type: string
          enum:
            - PERCENTAGE
            - FIXED_AMOUNT
        discount_value:
          type: number
          format: decimal
          minimum: 0.01
        min_order_amount:
          type: number
          format: decimal
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
          default: true

    PromoCodeResponse:
      type: object
      required:
        - id
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - current_uses
        - valid_from
        - valid_until
        - active
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discount_type:
          type: string
          enum:
            - PERCENTAGE
            - FIXED_AMOUNT
        discount_value:
          type: number
          format: decimal
        min_order_amount:
          type: number
          format: decimal
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean

    InsufficientStockDetails:
      type: object
      description: Детали при error_code INSUFFICIENT_STOCK
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - product_id
              - requested_quantity
              - available_stock
            properties:
              product_id:
                type: string
                format: uuid
              requested_quantity:
                type: integer
              available_stock:
                type: integer